# FamilyClub.React/famiy.club.react/lib/api/generated/models/LoginClubMemberDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginClubMemberDto
 */
export interface LoginClubMemberDto {
    /**
     * 
     * @type {string}
     * @memberof LoginClubMemberDto
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof LoginClubMemberDto
     */
    password: string;
    /**
     * 
     * @type {boolean}
     * @memberof LoginClubMemberDto
     */
    rememberMe?: boolean;
    /**
     * 
     * @type {string}
     * @memberof LoginClubMemberDto
     */
    returnUrl?: string | null;
}

/**
 * Check if a given object implements the LoginClubMemberDto interface.
 */
export function instanceOfLoginClubMemberDto(value: object): value is LoginClubMemberDto {
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function LoginClubMemberDtoFromJSON(json: any): LoginClubMemberDto {
    return LoginClubMemberDtoFromJSONTyped(json, false);
}

export function LoginClubMemberDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginClubMemberDto {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'],
        'password': json['password'],
        'rememberMe': json['rememberMe'] == null ? undefined : json['rememberMe'],
        'returnUrl': json['returnUrl'] == null ? undefined : json['returnUrl'],
    };
}

export function LoginClubMemberDtoToJSON(json: any): LoginClubMemberDto {
    return LoginClubMemberDtoToJSONTyped(json, false);
}

export function LoginClubMemberDtoToJSONTyped(value?: LoginClubMemberDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
        'rememberMe': value['rememberMe'],
        'returnUrl': value['returnUrl'],
    };
}