# FamilyClub.React/famiy.club.react/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FamilyClub.WebAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/AgeRestrictions": {
      "get": {
        "tags": [
          "AgeRestrictions"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AgeRestrictions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AgeRestrictions/{id}": {
      "get": {
        "tags": [
          "AgeRestrictions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AgeRestrictions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AgeRestrictions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AuthClubMember/register": {
      "post": {
        "tags": [
          "AuthClubMember"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.RegisterClubMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.RegisterClubMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.RegisterClubMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/AuthClubMember/login": {
      "post": {
        "tags": [
          "AuthClubMember"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.LoginClubMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.LoginClubMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.LoginClubMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.AuthResponseClubMemberDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.AuthResponseClubMemberDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.AuthResponseClubMemberDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Carts/{clubMemberId}": {
      "get": {
        "tags": [
          "Carts"
        ],
        "parameters": [
          {
            "name": "clubMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Cart.CartDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Cart.CartDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Cart.CartDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Carts"
        ],
        "parameters": [
          {
            "name": "clubMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Carts/{clubMemberId}/items": {
      "post": {
        "tags": [
          "Carts"
        ],
        "parameters": [
          {
            "name": "clubMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.CartItem.CartItemDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.CartItem.CartItemDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.CartItem.CartItemDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Carts/{clubMemberId}/items/{cartItemId}": {
      "put": {
        "tags": [
          "Carts"
        ],
        "parameters": [
          {
            "name": "clubMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cartItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Carts"
        ],
        "parameters": [
          {
            "name": "clubMemberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cartItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ClubMember/{id}": {
      "get": {
        "tags": [
          "ClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ClubMember"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FamilyClub.BLL.DTOs.AgeRestriction.AgeRestrictionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FamilyClub.BLL.DTOs.Cart.CartDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "clubMemberId": {
            "type": "string",
            "nullable": true
          },
          "cartItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.CartItem.CartItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyClub.BLL.DTOs.CartItem.CartItemDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cartId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "format": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "productPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FamilyClub.BLL.DTOs.ClubMember.AuthResponseClubMemberDTO": {
        "type": "object",
        "properties": {
          "clubMember": {
            "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiration": {
            "type": "string",
            "format": "date-time"
          },
          "returnUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FamilyClub.BLL.DTOs.ClubMember.ClubMemberReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "avatarData": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyClub.BLL.DTOs.ClubMember.LoginClubMemberDto": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "format": "password"
          },
          "rememberMe": {
            "type": "boolean"
          },
          "returnUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FamilyClub.BLL.DTOs.ClubMember.RegisterClubMemberDto": {
        "required": [
          "email",
          "password",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 6,
            "type": "string"
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string",
            "format": "tel"
          },
          "selectedRoles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}